修正在 XP 下無法將檔案解壓完成的問題。

+++ CrazyKTV_WebUpdater/MainWindow.xaml.cs
@@ -189,6 +189,13 @@ namespace CrazyKTV_WebUpdater
                 {
                     Process p = Process.Start(file, UnFolderFileArguments);
                     p.WaitForExit();
+                }
+            }
+
+            foreach (string file in FolderFileList)
+            {
+                if (File.Exists(file))
+                {
                     File.Delete(file);
                 }
             }
